pyproject.toml
[project]
name = "book-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "oleh.koval",email = "oleh.koval@nszu.gov.ua"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-genai (>=1.57.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "spacy (>=3.8.11,<4.0.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

[tool.poetry]
packages = [
    { include = "app", from = "src" }
]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["test"]
